+++ projects/demo/src/modules/pipes/field-error/examples/5/index.ts
@@ -44,7 +44,6 @@ export default class Example {
 
     constructor() {
         this.form = this.fb.group({
-            // eslint-disable-next-line i18n/no-russian-character
             text: ['русский текст', Validators.required],
         });